Scroll by half a screen instead of full screen
The former functionality of the scrollback patch is to scroll by n
lines, and for negative n's it will scroll by a screen minus n.

I don't use the negative functionality, and I like my vim behavior or
scrolling by half a screen with alt+u and alt+d, so I tweaked that to
reflect that behavior.

All negative n's will now yield the same result, which I'm ok with,
because I don't use the negative n functionality anyways.